refactor(core): extract window geometry and Engine constructor
Move streaming/long-form sample constants and the backend-aware
window_spec helpers into inference/windows so geometry stays testable
without a loaded model. Assemble Engine via from_loaded_parts so
load_with_factory no longer inlines the default attachment fields.

Constraint: no behavior change; decode seams and streaming stride stay bit-identical
Rejected: fold WindowSpec itself into windows.rs | it already lives with the audio window sources
Confidence: high
Scope-risk: narrow
